Fix critical app crash: TDZ-proof useApi, CNS error boundary, Dashboard HMR (#111-114)
#111 useApi.js fetchData TDZ crash:
- Converted fetchData from useCallback to useRef + stable wrapper pattern
- fetchDataRef.current assigned every render (always latest closure)
- fetchData wrapper has [] deps (never changes identity, no useEffect churn)
- Entire fetch body wrapped in try-catch (never throws unhandled)

#113 Graceful degradation when backend is down:
- CNSProvider: all 3 data-syncing useEffects wrapped in try-catch
- Layout.jsx: CNSErrorBoundary class component wraps CNSProvider
- LayoutFallback shows "System Initializing" with retry button and
  actionable "start backend" instructions instead of blank crash screen
- Auto-recovery: existing 5s retry loop + setError(null) on success

#114 Dashboard HMR failures:
- Root cause: 10 inline component definitions broke React Fast Refresh
- Extracted TickerStrip, HexagonLogo, RegimeDonut, TopTradesDonut,
  SignalBarChart, MiniEquityCurve, AgentConsensusRing, FlywheelPipeline,
  ConsensusBar, HeatmapGrid + SORT_PILLS/TIMEFRAMES to DashboardWidgets.jsx
- Dashboard.jsx: 2006 → 1234 lines (single component = reliable HMR)
